+++ spec/shoulda/matchers/active_model/context_matcher_spec.rb
@@ -1,10 +1,25 @@
 require 'spec_helper'
 
 describe Shoulda::Matchers::ActiveModel::ContextMatcher do
-  include ActiveModelHelpers::NumericalityHelpers
-
   context 'on' do
-    it { instance_with_validations(:on => :customisable).should matcher.on(:customisable) }
-    it { instance_without_validations.should_not matcher.on(:customisable) }
+    it { validating_something_in_context(:on => :customisable).should matcher.on(:customisable) }
+    it { validating_nothing.should_not matcher.on(:customisable) }
+  end
+
+  def validating_something_in_context(options = {})
+    define_model :example, :attr => :string do
+      validates_numericality_of :attr, options
+      attr_accessible :attr
+    end.new
+  end
+
+  def validating_nothing
+    define_model :example, :attr => :string do
+      attr_accessible :attr
+    end.new
+  end
+
+  def matcher
+    validate_numericality_of(:attr)
   end
 end
